+++ CHANGELOG.md
@@ -42,6 +42,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   supported.
 - Double arrow head types like `invdot` and `onormalonormal` once again display
   correctly. This was a regression in Graphviz 8.0.1. #2406
+- The `lvee` and `rvee` edge arrow shapes are slighty incorrect for
+  penwidths > 1. #2399
+- Small gap between `lcurve` or `rcurve` arrow shaft and node. #2426
+- Failure of arrowhead and arrowtail to respect penwidth #372 \
+  Fixed also for the `crow` and `vee`
+  [edge arrow shapes](https://graphviz.org/doc/info/arrows.html#primitive-shapes).
 
 ## [8.1.0] – 2023-07-06